Refactor the plaintext serving code into its own module
This is step one of many in this refactoring
